refactor(useradm): Make the http pkg structure more consistent
Move the endpoints to `router.go` and rename `middleware.go` to `utils.go`
